Refactors modules to obey private / public naming practices
